fix: update dockerfile to fix test (#84)
This gets tests we care about passing again, as Dean requested. It also comments out the arches we dont care about, and runs `go mod tidy`. 

I gave up on fixing depaware (there's some local <-> ci desync) as it doesn't seem particularly useful for us, it's not like we use it on coder/coder.

In any case, the coder/coder tests are the ones we really need to care about.